+++ packages/seeder/src/seeder.js
@@ -1,7 +1,6 @@
 const { EventEmitter } = require('events')
 const { homedir } = require('os')
 const { join } = require('path')
-const { promisify } = require('util')
 
 const Corestore = require('corestore')
 const HypercoreCache = require('hypercore-cache')
@@ -237,8 +236,6 @@ class Seeder extends EventEmitter {
 
     await this._networker.listen()
 
-    this._connectivity = promisify(this._networker.swarm.connectivity).bind(this._networker.swarm)
-
     const onPeerAdd = (peer) => {
       this.emit('networker-peer-add', {
         remoteAddress: peer.remoteAddress,
@@ -302,7 +299,9 @@ class Seeder extends EventEmitter {
   async getSwarmStats () {
     await this.init()
 
-    const { holepunched, bootstrapped } = await this._connectivity()
+    const holepunchable = this._networker.swarm.holepunchable()
+    const bootstrapped = holepunchable
+
     const ra = this._networker.swarm.remoteAddress()
     const remoteAddress = ra ? `${ra.host}:${ra.port}` : ''
     const currentPeers = Array
@@ -318,7 +317,7 @@ class Seeder extends EventEmitter {
       }, [])
 
     return {
-      holepunchable: holepunched,
+      holepunchable,
       bootstrapped,
       remoteAddress,
       currentPeers